Adapted varargs optimization test to the new varargs encoding.
